edit restuarant and owner role

+++ controllers/restaurants.js
@@ -99,6 +99,7 @@ exports.getRestaurant= async (req,res,next)=>{
 //Create new Restaurant 
 exports.createRestaurant = async(req,res,next)=>{
     try {
+        req.body.user = req.user.id;
         const restaurant = await Restaurant.create(req.body);
         
         res.status (201).json({
@@ -130,14 +131,23 @@ exports.updateRestaurant= async (req, res,next)=> {
         if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
             return res.status(400).json({success:false, message:'Invalid restaurant ID'});
         }
-        const restaurant = await Restaurant.findByIdAndUpdate (req.params.id, req.body, {
-            new: true,
-            runValidators:true
-        });
+        let restaurant = await Restaurant.findById(req.params.id);
 
         if(!restaurant){
             return res.status (404).json({success:false});
         }
+
+        // ตรวจสอบว่าผู้ใช้งานเป็นเจ้าของร้านนี้ หรือเป็น admin หรือไม่
+        if(restaurant.user.toString() !== req.user.id && req.user.role !== 'admin') {
+            return res.status(401).json({
+                success: false, 
+                message: `User ${req.user.id} is not authorized to update this restaurant`
+            });
+        }
+        restaurant = await Restaurant.findByIdAndUpdate (req.params.id, req.body, {
+            new: true,
+            runValidators:true
+        });
 
         res.status (200).json({success:true, data: restaurant});
     } catch (err) {
@@ -158,6 +168,14 @@ exports.deleteRestaurant = async (req,res,next)=>{
         {
             return res.status(404).json({success:false, message:`Restaurant not found with id of ${req.params.id}`});
         }
+        
+        if(restaurant.user.toString() !== req.user.id && req.user.role !== 'admin') {
+            return res.status(401).json({
+                success: false, 
+                message: `User ${req.user.id} is not authorized to delete this restaurant`
+            });
+        }
+
         await Reservation.deleteMany({ restaurant: req.params.id });
         await Restaurant.deleteOne({ _id: req.params.id });
